requirements.txt
customtkinter==5.2.0
pillow==10.1.0
pytest==7.4.3
python-dotenv==1.0.0
bcrypt==4.0.1
tkcalendar==1.6.1
matplotlib==3.8.0
reportlab==4.0.7
openpyxl==3.1.2